Applying (modified) patch to markup.py from ticket #6526 (http://code.djangoproject.com/attachment/ticket/6526/markup.py)
Adding the ability to override the default '[HTML_REMOVED]' message in the markup filter by putting a 'MARKDOWN_HTML_REMOVED_TEXT' in settings.py.

If the 'MARKDOWN_HTML_REMOVED_TEXT' is not set, it will default to markdown.HTML_REMOVED_TEXT which is '[HTML_REMOVED]'

Also updated docstring to reflect these changes.